+++ account/src/main/java/account/common/exception/UserExceptionHandler.java
@@ -1,12 +1,12 @@
 package account.common.exception;
 
 import global.api.Api;
+import global.errorcode.UserErrorCode;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import account.common.error.UserErrorCode;
 import account.common.exception.user.ExistUserEmailException;
 import account.common.exception.user.ExistUserNameException;
 import account.common.exception.user.LoginFailException;
