+++ board/src/main/java/board/common/exception/CommentExceptionHandler.java
@@ -1,8 +1,8 @@
 package board.common.exception;
 
-import board.common.error.CommentErrorCode;
 import board.common.exception.comment.CommentNotFoundException;
 import global.api.Api;
+import global.errorcode.CommentErrorCode;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
