+++ board/src/main/java/board/common/exception/UserExceptionHandler.java
@@ -1,12 +1,12 @@
 package board.common.exception;
 
 import global.api.Api;
+import global.errorcode.UserErrorCode;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import board.common.error.UserErrorCode;
 import board.common.exception.user.ExistUserEmailException;
 import board.common.exception.user.ExistUserNameException;
 import board.common.exception.user.LoginFailException;
