+++ global/src/main/java/global/errorcode/CommentErrorCode.java
@@ -1,6 +1,5 @@
-package board.common.error;
+package global.errorcode;
 
-import global.errorcode.ErrorCodeIfs;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
@@ -8,7 +7,8 @@ import lombok.Getter;
 @AllArgsConstructor
 public enum CommentErrorCode implements ErrorCodeIfs {
 
-    COMMENT_NOT_FOUND(404, 1400, "댓글이 존재하지 않습니다.")
+    // 1350 ~ 1399
+    COMMENT_NOT_FOUND(404, 1350, "댓글이 존재하지 않습니다.")
     ;
 
     private final Integer httpCode;
