+++ pet/src/main/java/pet/common/exception/PetExceptionHandler.java
@@ -1,12 +1,12 @@
 package pet.common.exception;
 
 import global.api.Api;
+import global.errorcode.PetErrorCode;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import pet.common.error.PetErrorCode;
 import pet.common.exception.pet.ExistsPetException;
 import pet.common.exception.pet.PetNotFoundException;
