+++ pet/src/main/java/pet/domain/pet/service/PetService.java
@@ -3,11 +3,11 @@ package pet.domain.pet.service;
 import db.domain.pet.PetEntity;
 import db.domain.pet.PetRepository;
 import db.domain.pet.enums.PetStatus;
+import global.errorcode.PetErrorCode;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import pet.common.error.PetErrorCode;
 import pet.common.exception.pet.ExistsPetException;
 import pet.common.exception.pet.PetNotFoundException;
 import pet.domain.pet.controller.model.register.PetRegisterRequest;
