+++ pet/src/main/java/pet/domain/user/interceptor/AuthorizationInterceptor.java
@@ -1,5 +1,6 @@
 package pet.domain.user.interceptor;
 
+import global.errorcode.UserErrorCode;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.Objects;
@@ -10,7 +11,6 @@ import org.springframework.web.context.request.RequestAttributes;
 import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.servlet.HandlerInterceptor;
 import org.springframework.web.servlet.resource.ResourceHttpRequestHandler;
-import pet.common.error.UserErrorCode;
 import pet.common.exception.user.UserNotFoundException;
 
 @Component
